requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2023.11.17
charset-normalizer==3.3.2
dgl==1.1.0+cu113
hydra-core==1.3.2
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.2
joblib==1.3.2
lightning-utilities==0.10.0
littleutils==0.2.2
MarkupSafe==2.1.3
networkx==3.1
numpy==1.24.4
ogb==1.3.6
omegaconf==2.3.0
outdated==0.2.2
packaging==23.2
pandas==2.0.3
psutil==5.9.7
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.2.0
torch==1.12.1+cu113
torch-scatter==2.1.0+pt112cu113
torch-sparse==0.6.16+pt112cu113
torch_geometric==2.4.0
torchmetrics==1.2.1
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
zipp==3.17.0
